refactor(api-client.ts): update import path for parseToolIdUtil to utils/tool-id.js refactor(openapi-loader.ts): update import path for abbreviations and generateToolId to utils folder refactor(tools-manager.ts): update import path for parseToolIdUtil to utils/tool-id.js feat(utils): add abbreviations and tool-id utility functions to improve code organization and reusability feat(utils): create tool-id utility functions for parsing and generating tool IDs
