+++ src/utils/abbreviations.ts
